[Refactor] Refactor CCE Addons create
