+++ openstack/cce/v3/addons/common.go
